Download Execucao de Despesas

Download Execucao de Despesas

download_exec_desp(year = NULL, month = NULL, ...)

Arguments

year
integer between 2013 and 2018

month
integer between 1 and 12

...
additional parameters

Value

data.frame

See also

Examples

(x <- download_exec_desp(year = 2014, month = 1))

#>        Ano e mes do lancamento Codigo Orgao Superior
#>     1:                 2014/01                 22000
#>     2:                 2014/01                 22000
#>     3:                 2014/01                 25000
#>     4:                 2014/01                 20000
#>     5:                 2014/01                 26000
#>    ---                                              
#> 51632:                 2014/01                 53000
#> 51633:                 2014/01                 26000
#> 51634:                 2014/01                 25000
#> 51635:                 2014/01                 26000
#> 51636:                 2014/01                 25000
#>                                 Nome Orgao Superior Codigo Orgao Subordinado
#>     1: MINIST. DA AGRICUL.,PECUARIA E ABASTECIMENTO                    22000
#>     2: MINIST. DA AGRICUL.,PECUARIA E ABASTECIMENTO                    22211
#>     3:                        MINISTERIO DA FAZENDA                    25000
#>     4:                     PRESIDENCIA DA REPUBLICA                    20101
#>     5:                       MINISTERIO DA EDUCACAO                    26410
#>    ---                                                                      
#> 51632:            MINISTERIO DA INTEGRACAO NACIONAL                    22203
#> 51633:                       MINISTERIO DA EDUCACAO                    26407
#> 51634:                        MINISTERIO DA FAZENDA                    25000
#> 51635:                       MINISTERIO DA EDUCACAO                    26247
#> 51636:                        MINISTERIO DA FAZENDA                    25000
#>                                                            Nome Orgao Subordinado
#>     1: MINIST. DA AGRICUL.,PECUARIA E ABASTECIMENTO - UNIDADES COM VINCULO DIRETO
#>     2:                                        COMPANHIA NACIONAL DE ABASTECIMENTO
#>     3:                        MINISTERIO DA FAZENDA - UNIDADES COM VINCULO DIRETO
#>     4:                                                   PRESIDENCIA DA REPUBLICA
#>     5:                               INST.FED.DE EDUC.,CIENC.E TEC.DO NORTE DE MG
#>    ---                                                                           
#> 51632:                              CIA.DE DES.DOS VALES DO S.FRANC.E DO PARNAIBA
#> 51633:                                       INST.FED.DE EDUC.,CIENC.E TEC.GOIANO
#> 51634:                        MINISTERIO DA FAZENDA - UNIDADES COM VINCULO DIRETO
#> 51635:                                        UNIVERSIDADE FEDERAL DE SANTA MARIA
#> 51636:                        MINISTERIO DA FAZENDA - UNIDADES COM VINCULO DIRETO
#>        Codigo Unidade Gestora                          Nome Unidade Gestora
#>     1:                 130069 SUPERINT.FED.DE AGRIC.,PEC.E ABASTECIMENTO/MA
#>     2:                 135476                      CONAB SEDE SUREG PARAIBA
#>     3:                 170055    SUPERINTENDeNCIA DE ADMINISTRAcaO DO MF/PE
#>     4:                 110001                SECRETARIA DE ADMINISTRACAO/PR
#>     5:                 158377        INST.FED.DO NORTE DE MG/CAMPUS SALINAS
#>    ---                                                                     
#> 51632:                 195002 CIA DE DES.DOS VALES DO S.FRANC.E DO PARNAIBA
#> 51633:                 158299                  IF GOIANO - CAMPUS RIO VERDE
#> 51634:                 170321                DELEGACIA DA RFB EM PIRACICABA
#> 51635:                 153164           UNIVERSIDADE FEDERAL DE SANTA MARIA
#> 51636:                 170142                  DELEGACIA DA RFB EM SOROCABA
#>        Codigo Gestao                          Nome Gestao
#>     1:             1                     TESOURO NACIONAL
#>     2:            NA                                     
#>     3:            NA                                     
#>     4:             1                     TESOURO NACIONAL
#>     5:            NA                                     
#>    ---                                                   
#> 51632:            NA                                     
#> 51633:         26407 INST.FED.DE EDUC.,CIENC.E TEC.GOIANO
#> 51634:             1                     TESOURO NACIONAL
#> 51635:            NA                                     
#> 51636:             1                     TESOURO NACIONAL
#>        Codigo Unidade Orcamentaria
#>     1:                       22101
#>     2:                       22211
#>     3:                       66101
#>     4:                       61101
#>     5:                       26410
#>    ---                            
#> 51632:                       53201
#> 51633:                       26407
#> 51634:                       25103
#> 51635:                       26247
#> 51636:                       25103
#>                            Nome Unidade Orcamentaria Codigo Funcao
#>     1:  MINISTERIO DA AGRIC.PECUARIA E ABASTECIMENTO            20
#>     2:           COMPANHIA NACIONAL DE ABASTECIMENTO            20
#>     3:         CONTROLADORIA-GERAL DA UNIAO - CGU/PR             4
#>     4:           SECRETARIA DE ASSUNTOS ESTRATEGICOS             4
#>     5:  INST.FED.DE EDUC.,CIENC.E TEC.DO NORTE DE MG            12
#>    ---                                                            
#> 51632: COMPANHIA DE DESENV. DO VALE DO SAO FRANCISCO            15
#> 51633:          INST.FED.DE EDUC.,CIENC.E TEC.GOIANO            12
#> 51634:                     RECEITA FEDERAL DO BRASIL             4
#> 51635:           UNIVERSIDADE FEDERAL DE SANTA MARIA            12
#> 51636:                     RECEITA FEDERAL DO BRASIL             4
#>          Nome Funcao Codigo Subfucao                       Nome Subfuncao
#>     1:   AGRICULTURA             125          NORMATIZACAO E FISCALIZACAO
#>     2:   AGRICULTURA             122                  ADMINISTRACAO GERAL
#>     3: ADMINISTRACAO             124                     CONTROLE INTERNO
#>     4: ADMINISTRACAO             331 PROTECAO E BENEFICIOS AO TRABALHADOR
#>     5:      EDUCACAO             363                  ENSINO PROFISSIONAL
#>    ---                                                                   
#> 51632:     URBANISMO             244              ASSISTENCIA COMUNITARIA
#> 51633:      EDUCACAO             363                  ENSINO PROFISSIONAL
#> 51634: ADMINISTRACAO             125          NORMATIZACAO E FISCALIZACAO
#> 51635:      EDUCACAO             364                      ENSINO SUPERIOR
#> 51636: ADMINISTRACAO             122                  ADMINISTRACAO GERAL
#>        Codigo Programa Orcamentario
#>     1:                         2028
#>     2:                         2105
#>     3:                         2101
#>     4:                         2101
#>     5:                         2031
#>    ---                             
#> 51632:                         2029
#> 51633:                         2031
#> 51634:                         2110
#> 51635:                         2032
#> 51636:                         2110
#>                                          Nome Programa Orcamentario Codigo Acao
#>     1:                                          DEFESA AGROPECUARIA        20ZX
#>     2: PROGRAMA DE GESTAO E MANUTENCAO DO MINISTERIO DA AGRICULTURA        2000
#>     3:  PROGRAMA DE GESTAO E MANUTENCAO DA PRESIDENCIA DA REPUBLICA        2D58
#>     4:  PROGRAMA DE GESTAO E MANUTENCAO DA PRESIDENCIA DA REPUBLICA        2012
#>     5:                          EDUCACAO PROFISSIONAL E TECNOLOGICA        20RL
#>    ---                                                                         
#> 51632:                       DESENVOLVIMENTO REGIONAL E TERRITORIAL        7K66
#> 51633:                          EDUCACAO PROFISSIONAL E TECNOLOGICA        20RL
#> 51634:     PROGRAMA DE GESTAO E MANUTENCAO DO MINISTERIO DA FAZENDA        2237
#> 51635: EDUCACAO SUPERIOR - GRADUACAO, POS-GRADUACAO, ENSINO, PESQUI        20RK
#> 51636:     PROGRAMA DE GESTAO E MANUTENCAO DO MINISTERIO DA FAZENDA        2000
#>                                                                            Nome Acao
#>     1:                                      FISCALIZACAO DE ATIVIDADES AGROPECUARIAS
#>     2:                                                      ADMINISTRACAO DA UNIDADE
#>     3:                CONTROLE INTERNO, PREVENCAO A CORRUPCAO, OUVIDORIA E CORREICAO
#>     4:              AUXILIO-ALIMENTACAO AOS SERVIDORES CIVIS, EMPREGADOS E MILITARES
#>     5: FUNCIONAMENTO DE INSTITUICOES FEDERAIS DE EDUCACAO PROFISSIONAL E TECNOLOGICA
#>    ---                                                                              
#> 51632:               APOIO A PROJETOS DE DESENVOLVIMENTO SUSTENTAVEL LOCAL INTEGRADO
#> 51633: FUNCIONAMENTO DE INSTITUICOES FEDERAIS DE EDUCACAO PROFISSIONAL E TECNOLOGICA
#> 51634:                               AUDITORIA E FISCALIZACAO TRIBUTARIA E ADUANEIRA
#> 51635:                     FUNCIONAMENTO DE INSTITUICOES FEDERAIS DE ENSINO SUPERIOR
#> 51636:                                                      ADMINISTRACAO DA UNIDADE
#>        Codigo Programa Governo Nome Programa Governo Codigo Grupo de Despesa
#>     1:                       0         NAO ATRIBUIDO                       3
#>     2:                      NA                                             3
#>     3:                      NA                                             3
#>     4:                       0         NAO ATRIBUIDO                       3
#>     5:                      NA                                             3
#>    ---                                                                      
#> 51632:                      NA                                             4
#> 51633:                       0         NAO ATRIBUIDO                       3
#> 51634:                       0         NAO ATRIBUIDO                       3
#> 51635:                      NA                                             3
#> 51636:                       0         NAO ATRIBUIDO                       3
#>            Nome Grupo de Despesa Codigo Elemento de Despesa
#>     1: OUTRAS DESPESAS CORRENTES                         30
#>     2: OUTRAS DESPESAS CORRENTES                         47
#>     3: OUTRAS DESPESAS CORRENTES                         30
#>     4: OUTRAS DESPESAS CORRENTES                         46
#>     5: OUTRAS DESPESAS CORRENTES                         20
#>    ---                                                     
#> 51632:             INVESTIMENTOS                         52
#> 51633: OUTRAS DESPESAS CORRENTES                         36
#> 51634: OUTRAS DESPESAS CORRENTES                         33
#> 51635: OUTRAS DESPESAS CORRENTES                         36
#> 51636: OUTRAS DESPESAS CORRENTES                         92
#>                            Nome Elemento de Despesa
#>     1:                          MATERIAL DE CONSUMO
#>     2:       OBRIGACOES TRIBUTARIAS E CONTRIBUTIVAS
#>     3:                          MATERIAL DE CONSUMO
#>     4:                          AUXILIO-ALIMENTACAO
#>     5:           AUXILIO FINANCEIRO A PESQUISADORES
#>    ---                                             
#> 51632:            EQUIPAMENTO E MATERIAL PERMANENTE
#> 51633: OUTROS SERVICOS DE TERCEIROS - PESSOA FISICA
#> 51634:           PASSAGENS E DESPESAS COM LOCOMOCAO
#> 51635: OUTROS SERVICOS DE TERCEIROS - PESSOA FISICA
#> 51636:            DESPESAS DE EXERCICIOS ANTERIORES
#>        Codigo Modalidade da Despesa Modalidade da Despesa Valor Empenhado (R$)
#>     1:                           90    APLICACOES DIRETAS              3000.00
#>     2:                           90    APLICACOES DIRETAS                 0.00
#>     3:                           90    APLICACOES DIRETAS                 0.00
#>     4:                           90    APLICACOES DIRETAS            214848.00
#>     5:                           90    APLICACOES DIRETAS                 0.00
#>    ---                                                                        
#> 51632:                           90    APLICACOES DIRETAS                 0.00
#> 51633:                           90    APLICACOES DIRETAS              1799.91
#> 51634:                           90    APLICACOES DIRETAS              1288.50
#> 51635:                           90    APLICACOES DIRETAS                 0.00
#> 51636:                           90    APLICACOES DIRETAS               107.08
#>        Valor Liquidado (R$) Valor Pago (R$) Valor Restos a Pagar Inscritos (R$)
#>     1:              3000.00         3000.00                                   0
#>     2:                 0.00            0.00                                   0
#>     3:                 0.00            0.00                                   0
#>     4:             27014.93        27014.93                                   0
#>     5:                 0.00            0.00                                   0
#>    ---                                                                         
#> 51632:                 0.00            0.00                                   0
#> 51633:              1799.91         1799.91                                   0
#> 51634:                 0.00            0.00                                   0
#> 51635:                 0.00            0.00                                   0
#> 51636:               107.08          107.08                                   0
#>        Valor Restos a Pagar Cancelado (R$) Valor Restos a Pagar Pagos (R$)
#>     1:                                   0                             0.0
#>     2:                                   0                           615.4
#>     3:                                   0                           249.0
#>     4:                                   0                             0.0
#>     5:                                   0                          2400.0
#>    ---                                                                    
#> 51632:                                   0                        581828.2
#> 51633:                                   0                             0.0
#> 51634:                                   0                             0.0
#> 51635:                                   0                          7935.0
#> 51636:                                   0                             0.0

(x <- download_exec_desp(year = 2014:2015, month = 1))

#>        Ano e mes do lancamento Codigo Orgao Superior
#>     1:                 2014/01                 22000
#>     2:                 2014/01                 22000
#>     3:                 2014/01                 25000
#>     4:                 2014/01                 20000
#>     5:                 2014/01                 26000
#>    ---                                              
#> 88650:                 2015/01                 22000
#> 88651:                 2015/01                 52000
#> 88652:                 2015/01                 26000
#> 88653:                 2015/01                 24000
#> 88654:                 2015/01                 25000
#>                                  Nome Orgao Superior Codigo Orgao Subordinado
#>     1:  MINIST. DA AGRICUL.,PECUARIA E ABASTECIMENTO                    22000
#>     2:  MINIST. DA AGRICUL.,PECUARIA E ABASTECIMENTO                    22211
#>     3:                         MINISTERIO DA FAZENDA                    25000
#>     4:                      PRESIDENCIA DA REPUBLICA                    20101
#>     5:                        MINISTERIO DA EDUCACAO                    26410
#>    ---                                                                       
#> 88650:  MINIST. DA AGRICUL.,PECUARIA E ABASTECIMENTO                    22211
#> 88651:                          MINISTERIO DA DEFESA                    52221
#> 88652:                        MINISTERIO DA EDUCACAO                    26238
#> 88653: MINIST.DA CIENCIA,TECNOL.,INOV.E COMUNICACOES                    24000
#> 88654:                         MINISTERIO DA FAZENDA                    25000
#>                                                             Nome Orgao Subordinado
#>     1:  MINIST. DA AGRICUL.,PECUARIA E ABASTECIMENTO - UNIDADES COM VINCULO DIRETO
#>     2:                                         COMPANHIA NACIONAL DE ABASTECIMENTO
#>     3:                         MINISTERIO DA FAZENDA - UNIDADES COM VINCULO DIRETO
#>     4:                                                    PRESIDENCIA DA REPUBLICA
#>     5:                                INST.FED.DE EDUC.,CIENC.E TEC.DO NORTE DE MG
#>    ---                                                                            
#> 88650:                                         COMPANHIA NACIONAL DE ABASTECIMENTO
#> 88651:                                      INDUSTRIA DE MATERIAL BELICO DO BRASIL
#> 88652:                                        UNIVERSIDADE FEDERAL DE MINAS GERAIS
#> 88653: MINIST.DA CIENCIA,TECNOL.,INOV.E COMUNICACOES - UNIDADES COM VINCULO DIRETO
#> 88654:                         MINISTERIO DA FAZENDA - UNIDADES COM VINCULO DIRETO
#>        Codigo Unidade Gestora                          Nome Unidade Gestora
#>     1:                 130069 SUPERINT.FED.DE AGRIC.,PEC.E ABASTECIMENTO/MA
#>     2:                 135476                      CONAB SEDE SUREG PARAIBA
#>     3:                 170055    SUPERINTENDeNCIA DE ADMINISTRAcaO DO MF/PE
#>     4:                 110001                SECRETARIA DE ADMINISTRACAO/PR
#>     5:                 158377        INST.FED.DO NORTE DE MG/CAMPUS SALINAS
#>    ---                                                                     
#> 88650:                 135476                      CONAB SEDE SUREG PARAIBA
#> 88651:                 168007   INDUSTRIA DE MATERIAL BELICO DO BRASIL/FMCE
#> 88652:                 153062          UNIVERSIDADE FEDERAL DE MINAS GERAIS
#> 88653:                 240127                  CENTRO DE TECNOLOGIA MINERAL
#> 88654:                 170072                   DELEGACIA DA RFB EM ARACAJU
#>        Codigo Gestao                            Nome Gestao
#>     1:             1                       TESOURO NACIONAL
#>     2:            NA                                       
#>     3:            NA                                       
#>     4:             1                       TESOURO NACIONAL
#>     5:            NA                                       
#>    ---                                                     
#> 88650:         22211    COMPANHIA NACIONAL DE ABASTECIMENTO
#> 88651:         16501 INDUSTRIA DE MATERIAL BELICO DO BRASIL
#> 88652:         15229   UNIVERSIDADE FEDERAL DE MINAS GERAIS
#> 88653:             1                       TESOURO NACIONAL
#> 88654:             1                       TESOURO NACIONAL
#>        Codigo Unidade Orcamentaria
#>     1:                       22101
#>     2:                       22211
#>     3:                       66101
#>     4:                       61101
#>     5:                       26410
#>    ---                            
#> 88650:                       22211
#> 88651:                       52221
#> 88652:                       26238
#> 88653:                       24101
#> 88654:                       25103
#>                            Nome Unidade Orcamentaria Codigo Funcao
#>     1:  MINISTERIO DA AGRIC.PECUARIA E ABASTECIMENTO            20
#>     2:           COMPANHIA NACIONAL DE ABASTECIMENTO            20
#>     3:         CONTROLADORIA-GERAL DA UNIAO - CGU/PR             4
#>     4:           SECRETARIA DE ASSUNTOS ESTRATEGICOS             4
#>     5:  INST.FED.DE EDUC.,CIENC.E TEC.DO NORTE DE MG            12
#>    ---                                                            
#> 88650:           COMPANHIA NACIONAL DE ABASTECIMENTO            20
#> 88651:  INDUSTRIA DE MATERIAL BELICO DO BRASIL-IMBEL             5
#> 88652:          UNIVERSIDADE FEDERAL DE MINAS GERAIS             9
#> 88653: MINIST.DA CIENCIA,TECNOL.,INOV.E COMUNICACOES            19
#> 88654:                     RECEITA FEDERAL DO BRASIL             4
#>                 Nome Funcao Codigo Subfucao
#>     1:          AGRICULTURA             125
#>     2:          AGRICULTURA             122
#>     3:        ADMINISTRACAO             124
#>     4:        ADMINISTRACAO             331
#>     5:             EDUCACAO             363
#>    ---                                     
#> 88650:          AGRICULTURA             122
#> 88651:      DEFESA NACIONAL             301
#> 88652:   PREVIDENCIA SOCIAL             272
#> 88653: CIENCIA E TECNOLOGIA             122
#> 88654:        ADMINISTRACAO             125
#>                              Nome Subfuncao Codigo Programa Orcamentario
#>     1:          NORMATIZACAO E FISCALIZACAO                         2028
#>     2:                  ADMINISTRACAO GERAL                         2105
#>     3:                     CONTROLE INTERNO                         2101
#>     4: PROTECAO E BENEFICIOS AO TRABALHADOR                         2101
#>     5:                  ENSINO PROFISSIONAL                         2031
#>    ---                                                                  
#> 88650:                  ADMINISTRACAO GERAL                         2105
#> 88651:                       ATENCAO BASICA                         2108
#> 88652:    PREVIDENCIA DO REGIME ESTATUTARIO                           89
#> 88653:                  ADMINISTRACAO GERAL                         2106
#> 88654:          NORMATIZACAO E FISCALIZACAO                         2110
#>                                          Nome Programa Orcamentario Codigo Acao
#>     1:                                          DEFESA AGROPECUARIA        20ZX
#>     2: PROGRAMA DE GESTAO E MANUTENCAO DO MINISTERIO DA AGRICULTURA        2000
#>     3:  PROGRAMA DE GESTAO E MANUTENCAO DA PRESIDENCIA DA REPUBLICA        2D58
#>     4:  PROGRAMA DE GESTAO E MANUTENCAO DA PRESIDENCIA DA REPUBLICA        2012
#>     5:                          EDUCACAO PROFISSIONAL E TECNOLOGICA        20RL
#>    ---                                                                         
#> 88650: PROGRAMA DE GESTAO E MANUTENCAO DO MINISTERIO DA AGRICULTURA        2000
#> 88651:      PROGRAMA DE GESTAO E MANUTENCAO DO MINISTERIO DA DEFESA        2004
#> 88652:              PREVIDENCIA DE INATIVOS E PENSIONISTAS DA UNIAO        0181
#> 88653: PROGRAMA DE GESTAO E MANUTENCAO DO MINISTERIO DA CIENCIA, TE        2000
#> 88654:     PROGRAMA DE GESTAO E MANUTENCAO DO MINISTERIO DA FAZENDA        2237
#>                                                                                               Nome Acao
#>     1:                                                         FISCALIZACAO DE ATIVIDADES AGROPECUARIAS
#>     2:                                                                         ADMINISTRACAO DA UNIDADE
#>     3:                                   CONTROLE INTERNO, PREVENCAO A CORRUPCAO, OUVIDORIA E CORREICAO
#>     4:                                 AUXILIO-ALIMENTACAO AOS SERVIDORES CIVIS, EMPREGADOS E MILITARES
#>     5:                    FUNCIONAMENTO DE INSTITUICOES FEDERAIS DE EDUCACAO PROFISSIONAL E TECNOLOGICA
#>    ---                                                                                                 
#> 88650:                                                                         ADMINISTRACAO DA UNIDADE
#> 88651: ASSISTENCIA MEDICA E ODONTOLOGICA AOS SERVIDORES CIVIS, EMPREGADOS, MILITARES E SEUS DEPENDENTES
#> 88652:                                                          APOSENTADORIAS E PENSOES CIVIS DA UNIAO
#> 88653:                                                                         ADMINISTRACAO DA UNIDADE
#> 88654:                                                  AUDITORIA E FISCALIZACAO TRIBUTARIA E ADUANEIRA
#>        Codigo Programa Governo Nome Programa Governo Codigo Grupo de Despesa
#>     1:                       0         NAO ATRIBUIDO                       3
#>     2:                      NA                                             3
#>     3:                      NA                                             3
#>     4:                       0         NAO ATRIBUIDO                       3
#>     5:                      NA                                             3
#>    ---                                                                      
#> 88650:                       0         NAO ATRIBUIDO                       3
#> 88651:                       0         NAO ATRIBUIDO                       3
#> 88652:                      -1        Sem informacao                       1
#> 88653:                       0         NAO ATRIBUIDO                       3
#> 88654:                       0         NAO ATRIBUIDO                       3
#>             Nome Grupo de Despesa Codigo Elemento de Despesa
#>     1:  OUTRAS DESPESAS CORRENTES                         30
#>     2:  OUTRAS DESPESAS CORRENTES                         47
#>     3:  OUTRAS DESPESAS CORRENTES                         30
#>     4:  OUTRAS DESPESAS CORRENTES                         46
#>     5:  OUTRAS DESPESAS CORRENTES                         20
#>    ---                                                      
#> 88650:  OUTRAS DESPESAS CORRENTES                         37
#> 88651:  OUTRAS DESPESAS CORRENTES                         39
#> 88652: PESSOAL E ENCARGOS SOCIAIS                         92
#> 88653:  OUTRAS DESPESAS CORRENTES                         37
#> 88654:  OUTRAS DESPESAS CORRENTES                         14
#>                            Nome Elemento de Despesa
#>     1:                          MATERIAL DE CONSUMO
#>     2:       OBRIGACOES TRIBUTARIAS E CONTRIBUTIVAS
#>     3:                          MATERIAL DE CONSUMO
#>     4:                          AUXILIO-ALIMENTACAO
#>     5:           AUXILIO FINANCEIRO A PESQUISADORES
#>    ---                                             
#> 88650:                       LOCACAO DE MAO-DE-OBRA
#> 88651: OUTROS SERVICOS DE TERCEIROS-PESSOA JURIDICA
#> 88652:            DESPESAS DE EXERCICIOS ANTERIORES
#> 88653:                       LOCACAO DE MAO-DE-OBRA
#> 88654:                      DIARIAS - PESSOAL CIVIL
#>        Codigo Modalidade da Despesa Modalidade da Despesa Valor Empenhado (R$)
#>     1:                           90    APLICACOES DIRETAS               3000.0
#>     2:                           90    APLICACOES DIRETAS                  0.0
#>     3:                           90    APLICACOES DIRETAS                  0.0
#>     4:                           90    APLICACOES DIRETAS             214848.0
#>     5:                           90    APLICACOES DIRETAS                  0.0
#>    ---                                                                        
#> 88650:                           90    APLICACOES DIRETAS                  0.0
#> 88651:                           90    APLICACOES DIRETAS               2000.0
#> 88652:                           90    APLICACOES DIRETAS             870102.0
#> 88653:                           90    APLICACOES DIRETAS             148567.8
#> 88654:                           90    APLICACOES DIRETAS                143.1
#>        Valor Liquidado (R$) Valor Pago (R$) Valor Restos a Pagar Inscritos (R$)
#>     1:              3000.00         3000.00                                   0
#>     2:                 0.00            0.00                                   0
#>     3:                 0.00            0.00                                   0
#>     4:             27014.93        27014.93                                   0
#>     5:                 0.00            0.00                                   0
#>    ---                                                                         
#> 88650:                 0.00            0.00                                   0
#> 88651:                 0.00            0.00                                   0
#> 88652:            224272.46       224272.46                                   0
#> 88653:                 0.00            0.00                                   0
#> 88654:               143.10          143.10                                   0
#>        Valor Restos a Pagar Cancelado (R$) Valor Restos a Pagar Pagos (R$)
#>     1:                                   0                            0.00
#>     2:                                   0                          615.40
#>     3:                                   0                          249.00
#>     4:                                   0                            0.00
#>     5:                                   0                         2400.00
#>    ---                                                                    
#> 88650:                                   0                        22086.96
#> 88651:                                   0                            0.00
#> 88652:                                   0                            0.00
#> 88653:                                   0                        38333.80
#> 88654:                                   0                            0.00

(x <- download_exec_desp(year = 2015, month = 1:2))

#>        Ano e mes do lancamento Codigo Orgao Superior
#>     1:                 2015/01                 26000
#>     2:                 2015/01                 52000
#>     3:                 2015/01                 25000
#>     4:                 2015/01                 26000
#>     5:                 2015/01                 26000
#>    ---                                              
#> 83717:                 2015/02                 25000
#> 83718:                 2015/02                 26000
#> 83719:                 2015/02                 53000
#> 83720:                 2015/02                 30000
#> 83721:                 2015/02                 36000
#>                              Nome Orgao Superior Codigo Orgao Subordinado
#>     1:                    MINISTERIO DA EDUCACAO                    26448
#>     2:                      MINISTERIO DA DEFESA                    52111
#>     3:                     MINISTERIO DA FAZENDA                    25207
#>     4:                    MINISTERIO DA EDUCACAO                    26402
#>     5:                    MINISTERIO DA EDUCACAO                    26424
#>    ---                                                                   
#> 83717:                     MINISTERIO DA FAZENDA                    25000
#> 83718:                    MINISTERIO DA EDUCACAO                    26420
#> 83719:         MINISTERIO DA INTEGRACAO NACIONAL                    22203
#> 83720: MINISTERIO DA JUSTICA E SEGURANcA PUBLICA                    30000
#> 83721:                       MINISTERIO DA SAUDE                    36000
#>                                                         Nome Orgao Subordinado
#>     1:                           UNIVERSIDADE FEDERAL DO SUL E SUDESTE DO PARA
#>     2:                                                  COMANDO DA AERONAUTICA
#>     3:                               SERVICO FEDERAL DE PROCESSAMENTO DE DADOS
#>     4:                                INST.FED.DE EDUC.,CIENC.E TEC.DE ALAGOAS
#>     5:                              INST.FED.DE EDUC.,CIENC.E TEC.DO TOCANTINS
#>    ---                                                                        
#> 83717:                     MINISTERIO DA FAZENDA - UNIDADES COM VINCULO DIRETO
#> 83718:                               INST.FED.DE EDUC.,CIENC.E TEC.FARROUPILHA
#> 83719:                           CIA.DE DES.DOS VALES DO S.FRANC.E DO PARNAIBA
#> 83720: MINISTERIO DA JUSTICA E SEGURANcA PUBLICA - UNIDADES COM VINCULO DIRETO
#> 83721:                       MINISTERIO DA SAUDE - UNIDADES COM VINCULO DIRETO
#>        Codigo Unidade Gestora                          Nome Unidade Gestora
#>     1:                 158718 UNIVERSIDADE FEDERAL DO SUL E SUDESTE DO PARA
#>     2:                 120016             GRUPAMENTO DE APOIO DE S J CAMPOS
#>     3:                 803080                   SERPRO - REGIONAL SAO PAULO
#>     4:                 158147      INST.FED.DE EDUC.,CIENC.E TEC.DE ALAGOAS
#>     5:                 158557   INST.FED.DO TOCANTINS/CAMPUS PORTO NACIONAL
#>    ---                                                                     
#> 83717:                 170114    SUPERINTENDeNCIA DE ADMINISTRAcaO DO MF/RJ
#> 83718:                 158127     INST.FED.DE EDUC.,CIENC.E TEC.FARROUPILHA
#> 83719:                 195005 CIA DE DES.DOS VALES DO S.FRANC.E DO PARNAIBA
#> 83720:                 200248 SECRETARIA EXTRAORD. DE SEG.P/GRANDES EVENTOS
#> 83721:                 250026                      NUCLEO ESTADUAL DO MS/PA
#>        Codigo Gestao                                   Nome Gestao
#>     1:         26448 UNIVERSIDADE FEDERAL DO SUL E SUDESTE DO PARA
#>     2:             1                              TESOURO NACIONAL
#>     3:         17205     SERVICO FEDERAL DE PROCESSAMENTO DE DADOS
#>     4:         26402      INST.FED.DE EDUC.,CIENC.E TEC.DE ALAGOAS
#>     5:         26424    INST.FED.DE EDUC.,CIENC.E TEC.DO TOCANTINS
#>    ---                                                            
#> 83717:             1                              TESOURO NACIONAL
#> 83718:         26420     INST.FED.DE EDUC.,CIENC.E TEC.FARROUPILHA
#> 83719:         11201 CIA DE DESENV DOS VALES DO S.FRANC.E PARNAIBA
#> 83720:             1                              TESOURO NACIONAL
#> 83721:             1                              TESOURO NACIONAL
#>        Codigo Unidade Orcamentaria
#>     1:                       26448
#>     2:                       24205
#>     3:                       25207
#>     4:                       26402
#>     5:                       26424
#>    ---                            
#> 83717:                       25104
#> 83718:                       26420
#> 83719:                       53201
#> 83720:                       30101
#> 83721:                       36901
#>                            Nome Unidade Orcamentaria Codigo Funcao
#>     1: UNIV.FED.DO SUL E SUDESTE DO PARA - UNIFESSPA            12
#>     2:             AGENCIA ESPACIAL BRASILEIRA - AEB            19
#>     3:     SERVICO FEDERAL DE PROCESSAMENTO DE DADOS             4
#>     4:      INST.FED.DE EDUC.,CIENC.E TEC.DE ALAGOAS            12
#>     5:    INST.FED.DE EDUC.,CIENC.E TEC.DO TOCANTINS            12
#>    ---                                                            
#> 83717:        PROCURADORIA GERAL DA FAZENDA NACIONAL             4
#> 83718:     INST.FED.DE EDUC.,CIENC.E TEC.FARROUPILHA            12
#> 83719: COMPANHIA DE DESENV. DO VALE DO SAO FRANCISCO            15
#> 83720:     MINISTERIO DA JUSTICA E SEGURANcA PUBLICA             6
#> 83721:                       FUNDO NACIONAL DE SAUDE            10
#>                 Nome Funcao Codigo Subfucao
#>     1:             EDUCACAO             331
#>     2: CIENCIA E TECNOLOGIA             572
#>     3:        ADMINISTRACAO             126
#>     4:             EDUCACAO             363
#>     5:             EDUCACAO             363
#>    ---                                     
#> 83717:        ADMINISTRACAO             122
#> 83718:             EDUCACAO             363
#> 83719:            URBANISMO             244
#> 83720:    SEGURANCA PUBLICA             122
#> 83721:                SAUDE             122
#>                                  Nome Subfuncao Codigo Programa Orcamentario
#>     1:     PROTECAO E BENEFICIOS AO TRABALHADOR                         2109
#>     2: DESENVOLVIMENTO TECNOLOGICO E ENGENHARIA                         2056
#>     3:                 TECNOLOGIA DA INFORMACAO                          751
#>     4:                      ENSINO PROFISSIONAL                         2109
#>     5:                      ENSINO PROFISSIONAL                         2031
#>    ---                                                                      
#> 83717:                      ADMINISTRACAO GERAL                         2110
#> 83718:                      ENSINO PROFISSIONAL                         2031
#> 83719:                  ASSISTENCIA COMUNITARIA                         2029
#> 83720:                      ADMINISTRACAO GERAL                         2112
#> 83721:                      ADMINISTRACAO GERAL                         2115
#>                                          Nome Programa Orcamentario Codigo Acao
#>     1:    PROGRAMA DE GESTAO E MANUTENCAO DO MINISTERIO DA EDUCACAO        2010
#>     2:                                            POLITICA ESPACIAL        20V0
#>     3:                 SERVICOS DE INFORMATICA PARA O SETOR PUBLICO        2211
#>     4:    PROGRAMA DE GESTAO E MANUTENCAO DO MINISTERIO DA EDUCACAO        20TP
#>     5:                          EDUCACAO PROFISSIONAL E TECNOLOGICA        20RL
#>    ---                                                                         
#> 83717:     PROGRAMA DE GESTAO E MANUTENCAO DO MINISTERIO DA FAZENDA        2000
#> 83718:                          EDUCACAO PROFISSIONAL E TECNOLOGICA        20RL
#> 83719:                       DESENVOLVIMENTO REGIONAL E TERRITORIAL        7K66
#> 83720: PROGRAMA DE GESTAO E MANUTENCAO DO MINISTERIO DA JUSTICA E S        2000
#> 83721:       PROGRAMA DE GESTAO E MANUTENCAO DO MINISTERIO DA SAUDE        2000
#>                                                                                                                         Nome Acao
#>     1:                                       ASSISTENCIA PRE-ESCOLAR AOS DEPENDENTES DOS SERVIDORES CIVIS, EMPREGADOS E MILITARES
#>     2: DESENVOLVIMENTO E LANCAMENTO DE FOGUETES SUBORBITAIS E DE VEICULOS LANCADORES DE SATELITES, COM A INFRAESTRUTURA ASSOCIADA
#>     3:                                                                SOLUcoES DE TECNOLOGIA DA INFORMAcaO PARA O GOVERNO FEDERAL
#>     4:                                                                                                      ATIVOS CIVIS DA UNIAO
#>     5:                                              FUNCIONAMENTO DE INSTITUICOES FEDERAIS DE EDUCACAO PROFISSIONAL E TECNOLOGICA
#>    ---                                                                                                                           
#> 83717:                                                                                                   ADMINISTRACAO DA UNIDADE
#> 83718:                                              FUNCIONAMENTO DE INSTITUICOES FEDERAIS DE EDUCACAO PROFISSIONAL E TECNOLOGICA
#> 83719:                                                            APOIO A PROJETOS DE DESENVOLVIMENTO SUSTENTAVEL LOCAL INTEGRADO
#> 83720:                                                                                                   ADMINISTRACAO DA UNIDADE
#> 83721:                                                                                                   ADMINISTRACAO DA UNIDADE
#>        Codigo Programa Governo Nome Programa Governo Codigo Grupo de Despesa
#>     1:                       0         NAO ATRIBUIDO                       3
#>     2:                       0         NAO ATRIBUIDO                       3
#>     3:                      -1        Sem informacao                       3
#>     4:                      -1        Sem informacao                       1
#>     5:                       0         NAO ATRIBUIDO                       3
#>    ---                                                                      
#> 83717:                       0         NAO ATRIBUIDO                       3
#> 83718:                       0         NAO ATRIBUIDO                       3
#> 83719:                       0         NAO ATRIBUIDO                       4
#> 83720:                      -1        Sem informacao                       3
#> 83721:                       0         NAO ATRIBUIDO                       3
#>             Nome Grupo de Despesa Codigo Elemento de Despesa
#>     1:  OUTRAS DESPESAS CORRENTES                          4
#>     2:  OUTRAS DESPESAS CORRENTES                         33
#>     3:  OUTRAS DESPESAS CORRENTES                         37
#>     4: PESSOAL E ENCARGOS SOCIAIS                         11
#>     5:  OUTRAS DESPESAS CORRENTES                         39
#>    ---                                                      
#> 83717:  OUTRAS DESPESAS CORRENTES                         14
#> 83718:  OUTRAS DESPESAS CORRENTES                         39
#> 83719:              INVESTIMENTOS                         47
#> 83720:  OUTRAS DESPESAS CORRENTES                         47
#> 83721:  OUTRAS DESPESAS CORRENTES                         36
#>                             Nome Elemento de Despesa
#>     1:             CONTRATACAO POR TEMPO DETERMINADO
#>     2:            PASSAGENS E DESPESAS COM LOCOMOCAO
#>     3:                        LOCACAO DE MAO-DE-OBRA
#>     4: VENCIMENTOS E VANTAGENS FIXAS - PESSOAL CIVIL
#>     5:  OUTROS SERVICOS DE TERCEIROS-PESSOA JURIDICA
#>    ---                                              
#> 83717:                       DIARIAS - PESSOAL CIVIL
#> 83718: OUTROS SERV.TERCEIROS-PES.JURID-OP.INTRA-ORC.
#> 83719:        OBRIGACOES TRIBUTARIAS E CONTRIBUTIVAS
#> 83720:        OBRIGACOES TRIBUTARIAS E CONTRIBUTIVAS
#> 83721:  OUTROS SERVICOS DE TERCEIROS - PESSOA FISICA
#>        Codigo Modalidade da Despesa
#>     1:                           90
#>     2:                           90
#>     3:                           90
#>     4:                           90
#>     5:                           90
#>    ---                             
#> 83717:                           90
#> 83718:                           91
#> 83719:                           90
#> 83720:                           90
#> 83721:                           90
#>                                Modalidade da Despesa Valor Empenhado (R$)
#>     1:                            APLICACOES DIRETAS               374.00
#>     2:                            APLICACOES DIRETAS                 0.00
#>     3:                            APLICACOES DIRETAS          11761720.12
#>     4:                            APLICACOES DIRETAS          10853280.40
#>     5:                            APLICACOES DIRETAS             33500.00
#>    ---                                                                   
#> 83717:                            APLICACOES DIRETAS               998.06
#> 83718: APLICACOES DIRETAS - OPER.INTRA-ORCAMENTARIAS              9370.32
#> 83719:                            APLICACOES DIRETAS                 0.00
#> 83720:                            APLICACOES DIRETAS              2458.00
#> 83721:                            APLICACOES DIRETAS             26767.72
#>        Valor Liquidado (R$) Valor Pago (R$) Valor Restos a Pagar Inscritos (R$)
#>     1:                62.70           62.70                                   0
#>     2:                 0.00            0.00                                   0
#>     3:             13985.76        10845.97                                   0
#>     4:          10787571.33     10787571.33                                   0
#>     5:             25766.76        21770.08                                   0
#>    ---                                                                         
#> 83717:              1069.61         1069.61                                   0
#> 83718:             29370.32        18450.32                                   0
#> 83719:                 0.00            0.00                                   0
#> 83720:              2238.00         2238.00                                   0
#> 83721:            108576.27       108576.27                                   0
#>        Valor Restos a Pagar Cancelado (R$) Valor Restos a Pagar Pagos (R$)
#>     1:                                   0                           62.70
#>     2:                                   0                         4484.88
#>     3:                                   0                       947142.63
#>     4:                                   0                         3250.65
#>     5:                                   0                        30809.08
#>    ---                                                                    
#> 83717:                                   0                            0.00
#> 83718:                                   0                        23354.53
#> 83719:                                   0                           26.38
#> 83720:                                   0                            0.00
#> 83721:                                   0                            0.00

# NOT RUN {
 (x <- download_exec_desp(year = "2014", month = 2))
 (x <- download_exec_desp(year = 2014))
# }